Interfaz funcional con programa de psoc y tiempo de vuelo minusculo, locura belho

+++ 4 - Pendulo.cydsn/codegentemp/UART_1.c
@@ -28,15 +28,15 @@ uint8 UART_1_initVar = 0u;
 
 #if (UART_1_TX_INTERRUPT_ENABLED && UART_1_TX_ENABLED)
     volatile uint8 UART_1_txBuffer[UART_1_TX_BUFFER_SIZE];
-    volatile uint8 UART_1_txBufferRead = 0u;
-    uint8 UART_1_txBufferWrite = 0u;
+    volatile uint16 UART_1_txBufferRead = 0u;
+    uint16 UART_1_txBufferWrite = 0u;
 #endif /* (UART_1_TX_INTERRUPT_ENABLED && UART_1_TX_ENABLED) */
 
 #if (UART_1_RX_INTERRUPT_ENABLED && (UART_1_RX_ENABLED || UART_1_HD_ENABLED))
     uint8 UART_1_errorStatus = 0u;
     volatile uint8 UART_1_rxBuffer[UART_1_RX_BUFFER_SIZE];
-    volatile uint8 UART_1_rxBufferRead  = 0u;
-    volatile uint8 UART_1_rxBufferWrite = 0u;
+    volatile uint16 UART_1_rxBufferRead  = 0u;
+    volatile uint16 UART_1_rxBufferWrite = 0u;
     volatile uint8 UART_1_rxBufferLoopDetect = 0u;
     volatile uint8 UART_1_rxBufferOverflow   = 0u;
     #if (UART_1_RXHW_ADDRESS_ENABLED)
@@ -395,8 +395,8 @@ void  UART_1_WriteControlRegister(uint8 control)
 
     #if (UART_1_RX_INTERRUPT_ENABLED)
 
-        uint8 locRxBufferRead;
-        uint8 locRxBufferWrite;
+        uint16 locRxBufferRead;
+        uint16 locRxBufferWrite;
 
         /* Protect variables that could change on interrupt */
         UART_1_DisableRxInt();
@@ -535,8 +535,8 @@ void  UART_1_WriteControlRegister(uint8 control)
         uint8 rxStatus;
 
     #if (UART_1_RX_INTERRUPT_ENABLED)
-        uint8 locRxBufferRead;
-        uint8 locRxBufferWrite;
+        uint16 locRxBufferRead;
+        uint16 locRxBufferWrite;
 
         /* Protect variables that could change on interrupt */
         UART_1_DisableRxInt();
@@ -665,7 +665,7 @@ void  UART_1_WriteControlRegister(uint8 control)
     *  None.
     *
     * Return:
-    *  uint8: Number of bytes in the RX buffer. 
+    *  uint16: Number of bytes in the RX buffer. 
     *    Return value type depends on RX Buffer Size parameter.
     *
     * Global Variables:
@@ -680,10 +680,10 @@ void  UART_1_WriteControlRegister(uint8 control)
     *  Allows the user to find out how full the RX Buffer is.
     *
     *******************************************************************************/
-    uint8 UART_1_GetRxBufferSize(void)
+    uint16 UART_1_GetRxBufferSize(void)
                                                             
     {
-        uint8 size;
+        uint16 size;
 
     #if (UART_1_RX_INTERRUPT_ENABLED)
 
@@ -1045,8 +1045,8 @@ void  UART_1_WriteControlRegister(uint8 control)
         *  to increment with a wrap, and we can't risk doing that with the real
         *  pointer and getting an interrupt in between instructions.
         */
-        uint8 locTxBufferWrite;
-        uint8 locTxBufferRead;
+        uint16 locTxBufferWrite;
+        uint16 locTxBufferRead;
 
         do
         { /* Block if software buffer is full, so we don't overwrite. */
@@ -1066,7 +1066,7 @@ void  UART_1_WriteControlRegister(uint8 control)
         }
         while( (locTxBufferWrite < locTxBufferRead) ? (locTxBufferWrite == (locTxBufferRead - 1u)) :
                                 ((locTxBufferWrite - locTxBufferRead) ==
-                                (uint8)(UART_1_TX_BUFFER_SIZE - 1u)) );
+                                (uint16)(UART_1_TX_BUFFER_SIZE - 1u)) );
 
         if( (locTxBufferRead == locTxBufferWrite) &&
             ((UART_1_TXSTATUS_REG & UART_1_TX_STS_FIFO_FULL) == 0u) )
@@ -1188,10 +1188,10 @@ void  UART_1_WriteControlRegister(uint8 control)
     *  TX buffer.
     *
     *******************************************************************************/
-    void UART_1_PutArray(const uint8 string[], uint8 byteCount)
+    void UART_1_PutArray(const uint8 string[], uint16 byteCount)
                                                                     
     {
-        uint8 bufIndex = 0u;
+        uint16 bufIndex = 0u;
 
         /* If not Initialized then skip this function */
         if(UART_1_initVar != 0u)
@@ -1270,10 +1270,10 @@ void  UART_1_WriteControlRegister(uint8 control)
     *  Allows the user to find out how full the TX Buffer is.
     *
     *******************************************************************************/
-    uint8 UART_1_GetTxBufferSize(void)
+    uint16 UART_1_GetTxBufferSize(void)
                                                             
     {
-        uint8 size;
+        uint16 size;
 
     #if (UART_1_TX_INTERRUPT_ENABLED)
